Extracted to a separate file for visibility
